requirement.txt
streamlit==1.26.0
langchain>=0.0.220
farm-haystack==1.19.0 
qdrant-client==1.10.0
sentence-transformers==2.2.2 
transformers==4.31.0
torch>=2.0.0
unstructured[local-inference]==0.10.19
python-dotenv
openai>=0.27.0,<0.29.0
uvicorn
requests
fastapi
python-multipart
pydantic